Change TitlebarButton to use more standard states

+++ src/TitlebarButtonNS.h
@@ -14,7 +14,7 @@
     BOOL _held;
 }
 
-@property (assign) int state;
+@property (assign) NSInteger state;
 @property (nonatomic, copy) NSViewDrawRect drawIcon;
 
 @end
